pyproject.toml
[tool.poetry]
name = "slothy-bot"
version = "0.1.0"
description = ""
authors = ["Lana Remenyuk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.19.0"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
isort = "^6.0.1"
redis = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
